trying to add floating pt number supp

+++ main.c
@@ -10,7 +10,7 @@ int add(int *a , int *b)
 
 */
 
-double operate(int *a, int *b, char *op)
+double operate(double *a, double *b, char *op)
 {
     //function to calculate vals 
     switch (*op)
@@ -26,7 +26,7 @@ double operate(int *a, int *b, char *op)
             //check if division by 0 
             if (*b != 0)
             {
-                return (double)*a / *b; 
+                return *a / *b; 
             }
             else
             {
@@ -49,24 +49,26 @@ double operate(int *a, int *b, char *op)
 
 int main()
 {
-    int num1, num2; //declare variables for input
+    double num1, num2; //declare variables for input
     char operator; 
 
 
     /* 
     
     printf("enter the first number: "); 
-    scanf("%d", &num1); //store input in num1 
+    scanf("%lf", &num1); //store input in num1 
     printf("enter the second number: "); 
-    scanf("%d", &num2); 
+    scanf("%lf", &num2); 
     */
 
-   printf("enter the expression: "); 
-   scanf("%d %c %d", &num1, &operator, &num2 ); 
+    printf("enter the expression: "); 
+    scanf("%lf %c %lf", &num1, &operator, &num2);
+
+
 
    double result = operate(&num1, &num2, &operator); 
 
-   printf("the result of %d %c %d is: %.2f\n", num1, operator, num2, result);  
+   printf("the result of %.2f %c %.2f is: %.2f\n", num1, operator, num2, result);
